+++ kitty/debug_config.py
@@ -91,7 +91,7 @@ def compare_opts(opts: KittyOpts, global_shortcuts: dict[str, SingleKey] | None,
             continue
         val = getattr(opts, f)
         if isinstance(val, dict):
-            print(title(f'{f}:'))
+            print(f'{f}:')
             if f == 'symbol_map':
                 for k in sorted(val):
                     print(f'\tU+{k[0]:04x} - U+{k[1]:04x} → {val[k]}')