+++ kitty/state.h
@@ -70,7 +70,14 @@ typedef struct Options {
     float macos_thicken_font;
     WindowTitleIn macos_show_window_title_in;
     char *bell_path, *bell_theme;
-    float background_opacity, dim_opacity, scrollback_indicator_opacity;
+    float background_opacity, dim_opacity;
+
+    ScrollbarVisibilityPolicy scrollbar;
+    bool scrollbar_interactive, scrollbar_jump_on_click;
+    float scrollbar_width, scrollbar_radius, scrollbar_gap, scrollbar_min_handle_height, scrollbar_hitbox_expansion;
+    float scrollbar_hover_width, scrollbar_handle_opacity, scrollbar_track_opacity, scrollbar_track_hover_opacity;
+    color_type scrollbar_handle_color, scrollbar_track_color;
+
     float text_contrast, text_gamma_adjustment;
     bool text_old_gamma;
 
@@ -144,6 +151,9 @@ typedef struct WindowLogoRenderData {
 
 typedef struct {
     unsigned int left, top, right, bottom;
+    struct {
+        unsigned int left, top, right, bottom;
+    } spaces;
 } WindowGeometry;
 
 typedef struct WindowRenderData {
@@ -214,6 +224,13 @@ typedef struct Window {
         PendingClick *clicks;
         size_t num, capacity;
     } pending_clicks;
+    struct {
+        double thumb_top, thumb_bottom;
+        bool is_dragging;
+        double drag_start_y;
+        double drag_start_scrolled_by;
+        bool is_hovering;
+    } scrollbar;
 } Window;
 
 typedef struct BorderRect {